Add library.properties file
The Arduino IDE supports two library formats. The first is the 1.0 format in which the library source files are in the root library folder. The second is the 1.5 format in which the library source files are in the src subfolder of the library. Libraries in 1.5 format are required to have a library.properties file in the root library folder, which contains metadata. With this file in place, it is possible to use the popular installation technique of downloading via GitHub's Clone or download > Download ZIP, then the Arduino IDE's Sketch > Include Library > Add .ZIP library.